+++ nacc_attribute_deriver/attributes/nacc/modules/cross_module.py
@@ -10,7 +10,8 @@ from nacc_attribute_deriver.attributes.base.namespace import (
 from nacc_attribute_deriver.attributes.base.uds_namespace import (
     UDSNamespace,
 )
-from nacc_attribute_deriver.schema.errors import MissingRequiredError
+
+# from nacc_attribute_deriver.schema.errors import MissingRequiredError
 from nacc_attribute_deriver.symbol_table import SymbolTable
 from nacc_attribute_deriver.utils.date import (
     calculate_age,